backend/requirements.txt
# Core Django and API
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# File handling & Media
Pillow==10.1.0
django-storages==1.14.2

# Internationalization
django-modeltranslation==0.18.11

# API Documentation
drf-spectacular==0.26.5

# Background tasks
celery==5.3.4
django-celery-beat==2.5.0

# Utilities
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2

# Development
django-debug-toolbar==4.2.0

# Communication
twilio==8.10.3

# Data processing
pandas==2.1.3
numpy==1.25.2

# Multi-tenancy
django-tenant-schemas==1.10.0

# Validation
marshmallow==3.20.1
